Add option for attribute rel="me" in social links
Fediverse services such as Mastodon verify ownership of URLs by looking
at a link of the following shape in a web site:
<a href="https://your.mastodon.tld/@profile" rel="me">...</a>

See here for official documentation:
https://docs.joinmastodon.org/user/profile/#verification

This change adds the "rel_me" boolean option for elements in the
"socials" section. If set to "true", the attribute will be added to the
respective <a> link tag.