Add config option for logo
Minor, backwards-compatible change that sets the logo through a config
option. This is helpful, for example, if you want to use a logo that is
not a png.

This change will have no effect on a default installation, as it
replaces a hardcoded path with a variable pointing to the same place.

Signed-off-by: Brian Warner <brian@bdwarner.com>